Add "T_2M" to the variables being queried
Unfortunately temperature is not uniformly called "T" but the
temperature at 2 meters height is still called "T_2M". This is a bug in
the dataset which needs to be fixed eventually. Until then, the variable
is queried and the name changed after querying and before writing it to
the CSV file.