Return only one DataFrame for radiation and temperature
This means we have to resample every DataFrame to 30 minute intervals
before joining it with the next one. Fortunately this doesn't change the
temperature DataFrames at all, because they are already in 30 minute
intervals. We also have to manipulate the column headers directly when
generating the DataFrames. But saving it to CSV files is a lot easier
now.
I also dropped support for the "T_2M" variable again, because I need to
check where this is missing data.