Change for which time intervals data is downloaded
Instead of downloading a whole interval around a certain point in time
within each year, now we are downloading a shorter interval around a
certain point in time in a specific number of days around that point in
time for each year. Takes a lot longer because while we are downloading
less data, we are opening a lot more connections to the server.