Put DataFrame into single element list
It's a hack to make the DataFrame available as the local variable `df`
inside the generator expression. This is done by moving the expression
generating the DataFrame into a single element list and then immediately
looping over this list, thus effectively just giving the DataFrame
inside the list a name. That way, the DataFrame is available inside
operations done to transform him, which we'll need later.