Generalize "radiation_and_temperature.py" to multiple locations
In order to do so, we now download the data into a `MultiIndex`ed
`DataFrame`, with the `MultiIndex`es first level being the location
`(x, y)` pairs and the second level being the time intervals. Resampling
got a little more complicated, because we now have to resample one axis
without touching the other one. Fortunately, the ridiculously powerful
and generic `groupby`, `Grouper` combo can do this.
Of course we also have to extract the location key from the data and
insert it into the `MultiIndex` which means one more level to loop over.
We also now have to construct and empty `MultiIndex` containing an empty
`DatetimeIndex` so that resampling doesn't raise an error.
Last but not least, since we can potentially download a lot of CSV
files, these are now saved in a dedicated directory: "csvfiles".