Donwload result data with a single request
This uses the new `ranges` parameter to `feedinlib.open_FRED.Weather`
which isn't released yet. But using that, we can use a single request to
download the complete data from all time intervals of interest using
only a single request. This also makes converting the result to a
`DataFrame` a lot less complicated. We just have to be a bit more
careful when resampling the result.